transform: own module for array transformations
